test_task_ati/requirements.txt
pytest==4.0.2
allure-pytest==2.6.3
selenium==3.141.0
PyHamcrest==1.9.0